RHOAIENG-9230 Restructure InitSegment
